+++ src/test/python/MemBlock/tests/test_MemBlock_env_fixture.py
@@ -20,6 +20,7 @@ def test_api_MemBlock_env_create(env):
     assert env.mock_outer_buffer is not None
     assert env.mock_dcache_client is not None
     assert env.mock_csr is not None
+    assert env.mmu is not None
     assert env.csr_agent is not None
     assert env.commit_agent is not None
     assert env.lsq_agent is not None